Split bbinternal package
part 1: bbinternal remains the stuff shared between bazel and standard
Go, the stuff that modifies AST, parses & writes Go files, and creates a
main.go.

part 2: findpkg turns a list of user-input package names/globs into
Packages we know.
