+++ src/pkg/bb/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/pkg/bb/bbinternal",
+        "//src/pkg/bb/findpkg",
         "//src/pkg/golang",
         "@com_github_google_goterm//term",
         "@com_github_u_root_uio//cp",
